requirements.txt
elasticsearch==7.8.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
requests==2.28.1